Fix overlay issue when changing languages
The root cause is that the name is saved as the overlay. When changing languages, the "Default" gets localized.  So it's not matched. To fix this, and issues where a user deletes or renames an overlay, I just added checks to map to the new "default" name if it's not found.